test: class構造化 cube 成功

+++ src/gfx/mesh.h
@@ -21,7 +21,7 @@ struct Mesh {
 	bgfx::IndexBufferHandle ibh{};
 	uint32_t indexCount{};
 
-	void create(const std::vector<Vertex>& v, const std::vector<uint16_t>& i) {
+	void create(const std::vector<Vertex>& v, const std::vector<uint32_t>& i) {
 		bgfx::VertexLayout layout;
 		Vertex::init(layout);
 
@@ -31,7 +31,7 @@ struct Mesh {
 		);
 
 		ibh = bgfx::createIndexBuffer(
-			bgfx::makeRef(i.data(), i.size()*sizeof(uint16_t))
+			bgfx::makeRef(i.data(), i.size()*sizeof(uint32_t))
 		);
 
 		indexCount = (uint32_t)i.size();
